examples/blog-writer/pyproject.toml
[tool.poetry]
name = "blog-writer"
version = "0.1.0"
description = ""
authors = ["bellabe <bella.belgarokova@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
groq = "^0.18.0"
python-dotenv = "^1.0.1"
py-trees = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
